Extract ConfigLoader
