+++ lib/nanoc/base/repos.rb
@@ -2,5 +2,6 @@ require_relative 'repos/store'
 
 require_relative 'repos/checksum_store'
 require_relative 'repos/compiled_content_cache'
+require_relative 'repos/config_loader'
 require_relative 'repos/rule_memory_store'
 require_relative 'repos/site_loader'
